+++ src/Public.Api/BuildingUnit/BackOffice/BuildingUnitBackOfficerController-AttachAddress.cs
@@ -1,8 +1,10 @@
-namespace Public.Api.Parcel.BackOffice
+namespace Public.Api.BuildingUnit.BackOffice
 {
     using System.Threading;
     using System.Threading.Tasks;
     using Be.Vlaanderen.Basisregisters.Api.Exceptions;
+    using BuildingRegistry.Api.BackOffice.Abstractions.BuildingUnit.Requests;
+    using BuildingRegistry.Api.Legacy.Abstractions.BuildingUnit.Responses;
     using Common.Infrastructure;
     using Common.Infrastructure.Extensions;
     using Infrastructure;
@@ -10,35 +12,34 @@ namespace Public.Api.Parcel.BackOffice
     using Microsoft.AspNetCore.Http;
     using Microsoft.AspNetCore.Mvc;
     using Microsoft.AspNetCore.Mvc.Infrastructure;
-    using ParcelRegistry.Api.BackOffice.Abstractions.Requests;
-    using ParcelRegistry.Api.Legacy.Parcel.Responses;
     using RestSharp;
     using Swashbuckle.AspNetCore.Annotations;
     using Swashbuckle.AspNetCore.Filters;
     using ProblemDetails = Be.Vlaanderen.Basisregisters.BasicApiProblem.ProblemDetails;
 
-    public partial class ParcelBackOfficeController
+    public partial class BuildingUnitBackOfficeController
     {
-        public const string DetachAddressParcelRoute = "percelen/{objectId}/acties/adresontkoppelen";
+        public const string AttachAddressBuildingUnitRoute = "gebouweenheden/{objectId}/acties/adreskoppelen";
 
         /// <summary>
-        /// Ontkoppel een adres van een perceel.
+        /// Koppel een adres aan een gebouweenheid.
         /// </summary>
-        /// <param name="objectId">Objectidentificator van het perceel (CaPaKey waarbij forward slash `/` vervangen werd door koppelteken `-`).</param>
+        /// <param name="objectId">Identificator van de gebouweenheid.</param>
+        /// <param name="attachAddressToBuildingUnitRequest"></param>
         /// <param name="actionContextAccessor"></param>
         /// <param name="problemDetailsHelper"></param>
-        /// <param name="ifMatch">If-Match header met ETag van de laatst gekende versie van het perceel (optioneel).</param>
-        /// <param name="toggle"></param>
+        /// <param name="attachAddressBuildingUnitToggle"></param>
+        /// <param name="ifMatch">If-Match header met ETag van de laatst gekende versie van de gebouweenheid (optioneel).</param>
         /// <param name="cancellationToken"></param>
         /// <response code="202">Als het ticket succesvol is aangemaakt.</response>
         /// <response code="400">Als uw verzoek foutieve data bevat.</response>
-        /// <response code="404">Als het perceel niet gevonden kan worden.</response>
+        /// <response code="404">Als de gebouweenheid niet gevonden kan worden.</response>
         /// <response code="406">Als het gevraagde formaat niet beschikbaar is.</response>
         /// <response code="412">Als de If-Match header niet overeenkomt met de laatste ETag.</response>
         /// <response code="429">Als het aantal requests per seconde de limiet overschreven heeft.</response>
         /// <response code="500">Als er een interne fout is opgetreden.</response>
         /// <returns></returns>
-        [ApiOrder(ApiOrder.Parcel.Edit + 2)]
+        [ApiOrder(ApiOrder.BuildingUnit.Edit + 12)]
         [ProducesResponseType(StatusCodes.Status202Accepted)]
         [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
         [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
@@ -47,40 +48,41 @@ namespace Public.Api.Parcel.BackOffice
         [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
         [SwaggerResponseHeader(StatusCodes.Status202Accepted, "location", "string", "De URL van het aangemaakte ticket.")]
         [SwaggerResponseHeader(StatusCodes.Status202Accepted, "x-correlation-id", "string", "Correlatie identificator van de response.")]
+        [SwaggerRequestExample(typeof(AttachAddressToBuildingUnitRequestExamples), typeof(AttachAddressToBuildingUnitRequestExamples))]
         [SwaggerResponseExample(StatusCodes.Status400BadRequest, typeof(BadRequestResponseExamples))]
-        [SwaggerResponseExample(StatusCodes.Status404NotFound, typeof(ParcelNotFoundResponseExamples))]
+        [SwaggerResponseExample(StatusCodes.Status404NotFound, typeof(BuildingUnitNotFoundResponseExamples))]
         [SwaggerResponseExample(StatusCodes.Status412PreconditionFailed, typeof(PreconditionFailedResponseExamples))]
         [SwaggerResponseExample(StatusCodes.Status429TooManyRequests, typeof(TooManyRequestsResponseExamples))]
         [SwaggerResponseExample(StatusCodes.Status500InternalServerError, typeof(InternalServerErrorResponseExamples))]
-        [SwaggerRequestExample(typeof(DetachAddressRequest), typeof(DetachAddressRequestExamples))]
-        [SwaggerOperation(Description = "Ontkoppel een adresId van een perceel.")]
-        [HttpPost(DetachAddressParcelRoute, Name = nameof(DetachAddressParcel))]
-        public async Task<IActionResult> DetachAddressParcel(
-            [FromBody] DetachAddressRequest request,
-            [FromRoute] string objectId,
+        [SwaggerOperation(Description = "Koppel een adresId met status `voorgesteld` of `inGebruik` aan een gebouweenheid met status `gepland` of `gerealiseerd`.")]
+        [HttpPost(AttachAddressBuildingUnitRoute, Name = nameof(AttachAddressBuildingUnit))]
+        public async Task<IActionResult> AttachAddressBuildingUnit(
+            [FromRoute] int objectId,
+            [FromBody] AttachAddressToBuildingUnitRequest attachAddressToBuildingUnitRequest,
             [FromServices] IActionContextAccessor actionContextAccessor,
             [FromServices] ProblemDetailsHelper problemDetailsHelper,
-            [FromServices] DetachAddressParcelToggle toggle,
+            [FromServices] AttachAddressBuildingUnitToggle attachAddressBuildingUnitToggle,
             [FromHeader(Name = HeaderNames.IfMatch)] string? ifMatch,
             CancellationToken cancellationToken = default)
         {
-            if (!toggle.FeatureEnabled)
+            if (!attachAddressBuildingUnitToggle.FeatureEnabled)
             {
                 return NotFound();
             }
 
             var contentFormat = DetermineFormat(actionContextAccessor.ActionContext);
 
-            RestRequest BackendRequest() => CreateBackendRequestWithJsonBody(DetachAddressParcelRoute, request, Method.Post)
+            RestRequest BackendRequest() =>
+                CreateBackendRequestWithJsonBody(AttachAddressBuildingUnitRoute, attachAddressToBuildingUnitRequest, Method.Post)
                 .AddParameter("objectId", objectId, ParameterType.UrlSegment)
                 .AddHeaderIfMatch(HeaderNames.IfMatch, ifMatch);
 
             var value = await GetFromBackendWithBadRequestAsync(
-                    contentFormat.ContentType,
-                    BackendRequest,
-                    CreateDefaultHandleBadRequest(),
-                    problemDetailsHelper,
-                    cancellationToken: cancellationToken);
+                contentFormat.ContentType,
+                BackendRequest,
+                CreateDefaultHandleBadRequest(),
+                problemDetailsHelper,
+                cancellationToken: cancellationToken);
 
             return new BackendResponseResult(value, BackendResponseResultOptions.ForBackOffice());
         }