+++ src/Public.Api/Infrastructure/Configuration/FeatureToggleOptions.cs
@@ -40,6 +40,7 @@ namespace Public.Api.Infrastructure.Configuration
         public bool CorrectBuildingNotRealization { get; set; }
         public bool ChangeBuildingGeometryOutline { get; set; }
         public bool DemolishBuilding { get; set; }
+        public bool RemoveBuilding { get; set; }
         public bool ChangeGeometryBuilding { get; set; }
         public bool CorrectGeometryBuilding { get; set; }